% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.r
\name{DigitData}
\alias{DigitData}
\title{Point data on an image}
\usage{
DigitData(col = "red", type = "p", ...)
}
\arguments{
\item{col}{Color of the points}

\item{type}{Type of the points}

\item{...}{Other arguments for locator}
}
\value{
\item{data}{A list with the coordinates of the points}
}
\description{
This function waits for the user to click the points of which he wants the
coordinates
}

