% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitize-package.r
\docType{package}
\name{digitize-package}
\alias{digitize-package}
\title{digitize : a plot digitizer in R}
\description{
Get data from a graph by providing calibration points
}
\details{
The package provides one main function, `digitize`, which runs functions
that 1) Read the image in and calibrate it, and 2) Digitize the data.
The first step requires user input.
}

