% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.r
\name{ReadAndCal}
\alias{ReadAndCal}
\title{Read and calibrate a JPG graphic}
\usage{
ReadAndCal(fname)
}
\arguments{
\item{fname}{Filename of the graphic to read}
}
\value{
\item{calpoints}{List of the x and y coordinates of the calibration
points}
}
\description{
This function loads the graphical file passed in argument, and allows the
user to calibrate it using four points (x1,x2,y1,y2) by clicking the graph.
}

