% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.r
\name{ReadImg}
\alias{ReadImg}
\title{Read a JPG graphic}
\usage{
ReadImg(fname)
}
\arguments{
\item{fname}{Filename of the graphic to read}
}
\description{
This function loads the graphical file passed in argument to allow further
point designation, but without calibration
}

