% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Calibrate}
\alias{Calibrate}
\title{(deprecated) Digitize the data}
\usage{
Calibrate(data, calpoints, x1, x2, y1, y2)
}
\arguments{
\item{data}{output of `DigitData`}

\item{calpoints}{output of `ReadAndCal`}

\item{x1}{X-coordinate of the leftmost x point (corrected)}

\item{x2}{X-coordinate of the rightmost x point (corrected)}

\item{y1}{Y-coordinate of the lower y point (corrected)}

\item{y2}{Y-coordinate of the upper y point (corrected)}
}
\value{
`data` A data frame with the corrected coordinates of the points
}
\description{
(deprecated) Digitize the data
}
\details{
deprecated. This function corrects the data according to the
calibration information. Usage further explained at
http://lukemiller.org/index.php/2011/06/digitizing-data-from-old-plots-using-digitize/
}
\examples{
\dontrun{Calibrate(data,calpoints,x1,x2,y1,y2)}
}

