% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{DigitData}
\alias{DigitData}
\title{(deprecated) Mark the data on an image}
\usage{
DigitData(col = "red", type = "p", ...)
}
\arguments{
\item{col}{color of marker as in `par`}

\item{type}{shape of marker as in `par`}

\item{...}{other args for `locator`}
}
\value{
`data` A list with the coordinates of the points
}
\description{
(deprecated) Mark the data on an image
}
\details{
deprecated. This function waits for the user to click the points of
the coordinates. See `graphics::locator` for more. Usage explained at
http://lukemiller.org/index.php/2011/06/digitizing-data-from-old-plots-using-digitize/
}

