% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ReadAndCal}
\alias{ReadAndCal}
\title{(deprecated) Read image and calibrate}
\usage{
ReadAndCal(fname)
}
\arguments{
\item{fname}{Filename of the graphic to read}
}
\value{
`calpoints` List of the x and y coordinates of the calibration points
}
\description{
(deprecated) Read image and calibrate
}
\details{
deprecated. Called for side effect of user locating points. See
`graphics::locator` for more. Usage explained at
http://lukemiller.org/index.php/2011/06/digitizing-data-from-old-plots-using-digitize/
}
\examples{
\dontrun{ReadAndCal(fname)}
}

